requirements.txt
Flask==2.0.1
twilio==7.15.0
python-dotenv==0.19.0
gunicorn==20.1.0